refactor: maintenance read model
